Resolving bugs

+++ app/src/main/java/com/example/swaad/ApiRequests/DataClassRestaurantsItem.kt
@@ -2,7 +2,7 @@ package com.example.swaad.ApiRequests
 
 data class DataClassRestaurantsItem(
     val address: String,
-    val avg_rating: Int,
+    val avg_rating:Double,
     val id: Int,
     val no_of_ratings: Int,
     val phone: String,
